moved user management out into its own file
